Improve format registration * Less interleaving (json does not require importing protobuf) * Generic text renderer
